backend/services/digest_generator/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Brotli==1.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
frozenlist==1.7.0
g4f==0.5.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
multidict==6.4.4
nest-asyncio==1.6.0
openai==1.82.0
propcache==0.3.2
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
redis==6.1.0
requests==2.32.4
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.1
